Split some async implementation classes into files
